+++ CHANGELOG.md
@@ -24,6 +24,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [0.1.0] - 2025-05-12 - Vlad
 ### Changed
 - Filter message on text
+- Gracefully stop docker
 
 ## [0.1.0] - 2025-05-09 - Vlad
 ### Added